mia_campaign_engine/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.29.0          # Azure PostgreSQL (async — API)
psycopg2-binary==2.9.10  # Azure PostgreSQL (sync — Celery workers)
alembic==1.14.0

# ─── Task Queue ───────────────────────────────────────────────────────────────
celery[redis]==5.4.0
kombu>=5.3.5              # Fix Redis SSL with redis-py 5.x
redis==5.2.0
flower==2.0.1            # Celery monitoring UI

# ─── Image Processing ─────────────────────────────────────────────────────────
Pillow==11.0.0
numpy==2.1.3

# ─── Data Handling ────────────────────────────────────────────────────────────
pandas==2.2.3
openpyxl==3.1.5

# ─── Azure SDK ────────────────────────────────────────────────────────────────
azure-storage-blob==12.23.1
azure-identity==1.19.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
tqdm==4.67.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.27.2
psutil==6.1.0

# ─── Video (optional moviepy as fallback, ffmpeg via subprocess is primary) ───
# moviepy==2.0.0         # Only if needed for complex compositing
